+++ Target Environment/Server/src/ComputeParams.cpp
@@ -22,7 +22,7 @@ static std::array<double, 6> comp_par_helper(const std::array<double, 6>& points
 }
 
 TargetEnvironment::TargetDesc compute_params(
-  const std::size_t id,
+  const std::uint32_t id,
   const std::array<double, 6>& points_x, 
   const std::array<double, 6>& points_y, 
   const std::array<double, 6>& points_z)
